+++ .gitignore
@@ -1,11 +1,12 @@
 /bin/run_dmft
+/bin/run_dmft/cal_spectrum
 /build/maxent/
 /build/impurity_solver/
 /build/libraries/
 /build/projection_embeding
 /libraries/gsl-2.6/
 /src/maxent/maxent
-/src/impurity_solver/CTHYB-LG/
+/src/impurity_solver/PACS_cthyb/
 /src/impurity_solver/Rutgers/
 /src/projecting_embending/bin/projection_embeding
 /src/projecting_embending/build/Makefile
