+++ README
@@ -1,0 +1,9 @@
+-------------
+Prerequisites
+-------------
+
+To install this package, the following is needed:
+1. C++ compiler: Intel >= 18.0
+2. Fortran compiler
+3. System environment requirements: gcc>=6.3.0, python3, scipy, numpy, mpi4py
+4. Intel MKL
