+++ bin/cal_spectrum
@@ -1,0 +1,27 @@
+#!/bin/bash
+nodes=1
+num_threads=1
+
+if [ $# -eq 0 ];then
+  echo "Warrning: Arguments nodes and num_threads are not given, and you will run the job with 1 process and 1 thread!!!"
+fi
+
+while [ $# != 0 ]
+do
+  case $1 in
+  "-nodes")
+    shift
+    nodes=$1
+  ;;
+  "-num_threads")
+    shift
+    num_threads=$1
+  ;;
+  *)
+    echo "ERROR: option $1 does not exist"
+  esac
+
+  shift;
+done
+
+mpirun -n $nodes -env OMP_NUM_THREADS=$num_threads /home/quxin/softwares/DFT_plus_DMFT/build/projection_embeding -eva.spectrum 1
