+++ src/projecting_embending/src/solver/rutgers_cthyb.h
@@ -1,8 +1,8 @@
-//=================================================================
-//    Computer Physics Communications 215 (2017) 128–136,
-//    https://doi.org/10.1016/j.cpc.2017.01.003
+//================================================================================
+//    Phys. Rev. B, 2007,215:155113. https://doi.org/10.1103/PhysRevB.75.155113
+//    Phys. Rev. B, 2010,81:195107. https://doi.org/10.1103/PhysRevB.81.195107
 //    The energy unit used is eV
-//=================================================================
+//=================================================================================
 #pragma once
 
 #include "../para/KS_bands.h"
